Panzer: Setting Tpetra assembly/solve to work with long GOs only
Previously this was only working with int so we were effectively upper
bounded by 2 billion still. This changes that. One issue however
is that you cannot switch dynamically (or at all!) between Tpetra<int,long>
to Tpetra<int,int>. The unit tests all pass at this point. Still need
to fix Drekar and ensure that MueLu works with Tpetra<int,long> through
Stratimikos.

+++ packages/panzer/src/evaluators/Panzer_ScatterDirichletResidual_Tpetra.cpp
@@ -52,6 +52,7 @@
 #include "Panzer_ScatterDirichletResidual_TpetraSG_decl.hpp"
 #include "Panzer_ScatterDirichletResidual_TpetraSG_impl.hpp"
 
-PANZER_INSTANTIATE_TEMPLATE_CLASS_FOUR_T(panzer::ScatterDirichletResidual_Tpetra,int,int)
+// PANZER_INSTANTIATE_TEMPLATE_CLASS_FOUR_T(panzer::ScatterDirichletResidual_Tpetra,int,int)
+PANZER_INSTANTIATE_TEMPLATE_CLASS_FOUR_T(panzer::ScatterDirichletResidual_Tpetra,int,long)
 
 #endif
